Refactor for web-ifc-babylon

+++ web-ifc-babylon/src/loaders/IFC/components/IFCParser.ts
@@ -1,4 +1,4 @@
-//@ts-ignore
+import { Color3, Geometry, Material, Matrix, Mesh, Nullable, StandardMaterial, TransformNode, VertexData } from '@babylonjs/core';
 import {
     PlacedGeometry,
     Color as ifcColor,
@@ -9,17 +9,7 @@ import {
     IFCPRODUCTDEFINITIONSHAPE
 } from 'web-ifc';
 import { IfcState, IfcMesh } from '../BaseDefinitions';
-import {
-    Color,
-    MeshLambertMaterial,
-    DoubleSide,
-    Matrix4,
-    BufferGeometry,
-    BufferAttribute,
-    Mesh
-} from 'three';
-import { mergeBufferGeometries } from 'three/examples/jsm/utils/BufferGeometryUtils';
-import { BvhManager } from './BvhManager';
+
 import { IFCModel } from './IFCModel';
 
 export interface ParserProgress {
@@ -43,8 +33,8 @@ export interface ParserAPI {
 
 export interface GeometriesByMaterial {
     [materialID: string]: {
-        material: MeshLambertMaterial,
-        geometries: BufferGeometry[]
+        material: Material,
+        geometries: Mesh[]
     }
 }
 
@@ -59,7 +49,7 @@ export class IFCParser implements ParserAPI {
         [IFCOPENINGELEMENT]: false
     };
 
-    private geometriesByMaterials: GeometriesByMaterial = {};
+    private geometriesByMaterials: Record<number, Mesh[]> = {};
 
     private loadingState = {
         total: 0,
@@ -75,34 +65,36 @@ export class IFCParser implements ParserAPI {
 
     // BVH is optional because when using workers we have to apply it in the main thread,
     // once the model has been serialized and reconstructed
-    constructor(private state: IfcState, private BVH?: BvhManager) {
+    constructor(private state: IfcState) {
     }
 
     async setupOptionalCategories(config: OptionalCategories) {
         this.optionalCategories = config;
     }
 
-    async parse(buffer: any, coordinationMatrix?: number[]) {
-        if (this.state.api.wasmModule === undefined) await this.state.api.Init();
+    async parse(buffer: any, coordinationMatrix?: number[]): Promise<IFCModel> {
+
+        if (this.state.api?.wasmModule === undefined) await this.state.api?.Init();
         await this.newIfcModel(buffer);
+
         this.loadedModels++;
-        if (coordinationMatrix) {
-            await this.state.api.SetGeometryTransformation(this.currentWebIfcID, coordinationMatrix);
+        if (coordinationMatrix && coordinationMatrix.length === 16) {
+            await this.state.api?.SetGeometryTransformation(this.currentWebIfcID, coordinationMatrix);
         }
         return this.loadAllGeometry(this.currentWebIfcID);
     }
 
     getAndClearErrors(_modelId: number) {
-        // return this.state.api.GetAndClearErrors(modelId);
+        return this.state.api?.GetAndClearErrors(_modelId);
     }
 
     private notifyProgress(loaded: number, total: number) {
         if (this.state.onProgress) this.state.onProgress({ loaded, total });
     }
 
-    private async newIfcModel(buffer: any) {
+    private async newIfcModel(buffer: ArrayBuffer) {
         const data = new Uint8Array(buffer);
-        this.currentWebIfcID = await this.state.api.OpenModel(data, this.state.webIfcSettings);
+        this.currentWebIfcID = await this.state.api?.OpenModel(data, this.state.webIfcSettings)!;
         this.currentModelID = this.state.useJSON ? this.loadedModels : this.currentWebIfcID;
         this.state.models[this.currentModelID] = {
             modelID: this.currentModelID,
@@ -112,38 +104,36 @@ export class IFCParser implements ParserAPI {
         };
     }
 
-    private async loadAllGeometry(modelID: number) {
+    private async loadAllGeometry(modelID: number): Promise<IFCModel> {
+        const model = new IFCModel(modelID.toString());
+
         this.addOptionalCategories(modelID);
         await this.initializeLoadingState(modelID);
 
-        this.state.api.StreamAllMeshes(modelID, (mesh: FlatMesh) => {
+        this.state.api?.StreamAllMeshes(modelID, (mesh: FlatMesh) => {
             this.updateLoadingState();
             // only during the lifetime of this function call, the geometry is available in memory
             this.streamMesh(modelID, mesh);
         });
 
         this.notifyLoadingEnded();
-        const geometries: BufferGeometry[] = [];
-        const materials: MeshLambertMaterial[] = [];
 
-        Object.keys(this.geometriesByMaterials).forEach((key) => {
-            const geometriesByMaterial = this.geometriesByMaterials[key].geometries;
-            const merged = mergeBufferGeometries(geometriesByMaterial);
-            materials.push(this.geometriesByMaterials[key].material);
-            geometries.push(merged);
+        Object.values(this.geometriesByMaterials).forEach(async (geometries) => {
+
+            const geometry = await Mesh.MergeMeshesAsync(geometries, true, true) as Mesh;
+            geometry.freezeWorldMatrix();
+            geometry.parent = model;
         });
 
-        const combinedGeometry = mergeBufferGeometries(geometries, true);
-        this.cleanUpGeometryMemory(geometries);
-        if (this.BVH) this.BVH.applyThreeMeshBVH(combinedGeometry);
-        const model = new IFCModel(combinedGeometry, materials);
+        this.geometriesByMaterials = {};
+
         this.state.models[this.currentModelID].mesh = model;
         return model;
     }
 
     private async initializeLoadingState(modelID: number) {
-        const shapes = await this.state.api.GetLineIDsWithType(modelID, IFCPRODUCTDEFINITIONSHAPE);
-        this.loadingState.total = shapes.size();
+        const shapes = await this.state.api!.GetLineIDsWithType(modelID, IFCPRODUCTDEFINITIONSHAPE);
+        this.loadingState.total = shapes?.size();
         this.loadingState.current = 0;
         this.loadingState.step = 0.1;
     }
@@ -154,7 +144,7 @@ export class IFCParser implements ParserAPI {
 
     private updateLoadingState() {
         const realCurrentItem = Math.min(this.loadingState.current++, this.loadingState.total);
-        if(realCurrentItem / this.loadingState.total >= this.loadingState.step) {
+        if (realCurrentItem / this.loadingState.total >= this.loadingState.step) {
             const currentProgress = Math.ceil(this.loadingState.total * this.loadingState.step);
             this.notifyProgress(currentProgress, this.loadingState.total);
             this.loadingState.step += 0.1;
@@ -173,7 +163,7 @@ export class IFCParser implements ParserAPI {
             }
         }
 
-        this.state.api.StreamAllMeshesWithTypes(this.currentWebIfcID, optionalTypes, (mesh: FlatMesh) => {
+        this.state.api?.StreamAllMeshesWithTypes(this.currentWebIfcID, optionalTypes, (mesh: FlatMesh) => {
             this.streamMesh(modelID, mesh);
         });
     }
@@ -184,58 +174,64 @@ export class IFCParser implements ParserAPI {
 
         for (let i = 0; i < size; i++) {
             const placedGeometry = placedGeometries.get(i);
-            let itemMesh = this.getPlacedGeometry(modelID, mesh.expressID, placedGeometry);
-            let geom = itemMesh.geometry.applyMatrix4(itemMesh.matrix);
-            this.storeGeometryByMaterial(placedGeometry.color, geom);
+            const geometry = this.getPlacedGeometry(modelID, mesh.expressID, placedGeometry)!;
+            if (geometry !== null)
+                this.storeGeometryByMaterial(placedGeometry.color, geometry);
         }
     }
 
-    private getPlacedGeometry(modelID: number, expressID: number, placedGeometry: PlacedGeometry) {
+    private getPlacedGeometry(modelID: number, expressID: number, placedGeometry: PlacedGeometry): Nullable<Mesh> {
         const geometry = this.getBufferGeometry(modelID, expressID, placedGeometry);
-        const mesh = new Mesh(geometry);
-        mesh.matrix = this.getMeshMatrix(placedGeometry.flatTransformation);
-        mesh.matrixAutoUpdate = false;
-        return mesh;
+        if (geometry === null) return null;
+
+        const matrix = Matrix.FromArray(placedGeometry.flatTransformation);
+
+        // Some IFC files are not parsed correctly, leading to degenerated meshes
+        try {
+            geometry.bakeTransformIntoVertices(matrix);
+        }
+        catch {
+            console.warn("Unable to bake transform matrix into vertex array. Some elements may be in the wrong position.");
+        }
+
+        return geometry;
     }
 
-    private getBufferGeometry(modelID: number, expressID: number, placedGeometry: PlacedGeometry) {
-        const geometry = this.state.api.GetGeometry(modelID, placedGeometry.geometryExpressID) as IfcGeometry;
-        const verts = this.state.api.GetVertexArray(geometry.GetVertexData(), geometry.GetVertexDataSize()) as Float32Array;
-        const indices = this.state.api.GetIndexArray(geometry.GetIndexData(), geometry.GetIndexDataSize()) as Uint32Array;
+    private getBufferGeometry(modelID: number, expressID: number, placedGeometry: PlacedGeometry): Nullable<Mesh> {
+        // WARNING: geometry must be deleted when requested from WASM
+        const geometry = this.state.api?.GetGeometry(modelID, placedGeometry.geometryExpressID) as IfcGeometry;
+        if (geometry.GetVertexDataSize() === 0) return null;
+        const verts = this.state.api?.GetVertexArray(geometry.GetVertexData(), geometry.GetVertexDataSize()) as Float32Array;
+        const indices = this.state.api?.GetIndexArray(geometry.GetIndexData(), geometry.GetIndexDataSize()) as Uint32Array;
+
         const buffer = this.ifcGeometryToBuffer(expressID, verts, indices);
         //@ts-ignore
         geometry.delete();
         return buffer;
     }
 
-    private storeGeometryByMaterial(color: ifcColor, geometry: BufferGeometry) {
-        let colID = `${color.x}${color.y}${color.z}${color.w}`;
-        if (this.geometriesByMaterials[colID]) {
-            this.geometriesByMaterials[colID].geometries.push(geometry);
+    private storeGeometryByMaterial(color: ifcColor, geometry: Mesh) {
+
+        let colorid: number = Math.floor(color.x * 256) + Math.floor(color.y * 256 ** 2)
+            + Math.floor(color.z * 256 ** 3) + Math.floor(color.w * 256 ** 4);
+        if (this.geometriesByMaterials[colorid]) {
+            this.geometriesByMaterials[colorid].push(geometry);
             return;
         }
 
-        // Assume RGB components are in sRGB-Rec709-D65 colorspace, and specify
-        // this so three.js can convert if THREE.ColorManagement APIs are enabled.
-        const col = new Color().setRGB(color.x, color.y, color.z, 'srgb');
-        const material = new MeshLambertMaterial({ color: col, side: DoubleSide });
-        material.transparent = color.w !== 1;
-        if (material.transparent) material.opacity = color.w;
-        this.geometriesByMaterials[colID] = { material, geometries: [geometry] };
+        // Assume RGB components are in sRGB-Rec709-D65 colorspace
+        const newMaterial = this.getMeshMaterial(color);
+        newMaterial.freeze();
+        geometry.material = newMaterial;
+        this.geometriesByMaterials[colorid] = [geometry];
     }
 
-    private getMeshMatrix(matrix: Array<number>) {
-        const mat = new Matrix4();
-        mat.fromArray(matrix);
-        return mat;
-    }
-
-    private ifcGeometryToBuffer(expressID: number, vertexData: Float32Array, indexData: Uint32Array) {
-        const geometry = new BufferGeometry();
+    private ifcGeometryToBuffer(expressID: number, vertexData: Float32Array, indexData: Uint32Array): Mesh {
+        const geometry = new Mesh(expressID.toString());
+        geometry.freezeWorldMatrix();
 
         const posFloats = new Float32Array(vertexData.length / 2);
         const normFloats = new Float32Array(vertexData.length / 2);
-        const idAttribute = new Uint32Array(vertexData.length / 6);
 
         for (let i = 0; i < vertexData.length; i += 6) {
             posFloats[i / 2] = vertexData[i];
@@ -245,35 +241,28 @@ export class IFCParser implements ParserAPI {
             normFloats[i / 2] = vertexData[i + 3];
             normFloats[i / 2 + 1] = vertexData[i + 4];
             normFloats[i / 2 + 2] = vertexData[i + 5];
-
-            idAttribute[i / 6] = expressID;
         }
 
-        geometry.setAttribute(
-            'position',
-            new BufferAttribute(posFloats, 3));
-        geometry.setAttribute(
-            'normal',
-            new BufferAttribute(normFloats, 3));
-        geometry.setAttribute(
-            'expressID',
-            new BufferAttribute(idAttribute, 1));
+        const data = new VertexData();
+        data.positions = posFloats;
+        data.normals = normFloats;
+        data.indices = indexData;
 
-        geometry.setIndex(new BufferAttribute(indexData, 1));
+        data.applyToMesh(geometry);
+
         return geometry;
     }
 
-    // Three.js geometry has to be manually deallocated
-    private cleanUpGeometryMemory(geometries: BufferGeometry[]) {
-        geometries.forEach(geometry => geometry.dispose());
+    private getMeshMaterial(color: ifcColor): Material {
+        const name = `(${Math.floor(color.x * 256)},${Math.floor(color.y * 256)},${Math.floor(color.z * 256)},${Math.floor(color.w * 256)})`;
+        const material = new StandardMaterial(name);
+        material.emissiveColor = new Color3(color.x, color.y, color.z);
+        // if material has alpha - make it fully transparent for performance
+        material.alpha = (color.w < 1.0 ? 0 : 1);
+        material.sideOrientation = Mesh.DOUBLESIDE;
+        material.backFaceCulling = false;
+        material.disableLighting = true;
 
-        Object.keys(this.geometriesByMaterials).forEach((materialID) => {
-            const geometriesByMaterial = this.geometriesByMaterials[materialID];
-            geometriesByMaterial.geometries.forEach(geometry => geometry.dispose());
-            geometriesByMaterial.geometries = [];
-            // @ts-ignore
-            geometriesByMaterial.material = null;
-        });
-        this.geometriesByMaterials = {};
+        return material;
     }
 }
